+++ changelog_entry.yaml
@@ -1,4 +1,4 @@
 - bump: minor
   changes:
     added:
-    - Reafctor the Alternative Minimum Tax (AMT) logic.
+    - Refactor the Alternative Minimum Tax (AMT) logic.
